Move all built in actions to this system
