+++ vizro-core/src/vizro/actions/_on_page_load_action.py
@@ -13,34 +13,20 @@ from vizro.models._action._action import NewAction
 class _on_page_load(NewAction):
     targets: list[ModelID]
 
-    def function(
+    def actual_function(
         self,
-        filters: list[State],
-        parameters: list[State],
-        filter_interaction: list[dict[str, State]],
+        filters,
+        parameters,
+        filter_interaction,
     ) -> dict[ModelID, Any]:
-        """Applies controls to charts on page once the page is opened (or refreshed).
-
-        Args:
-            targets: List of target component ids to filter by chart interaction. If missing, will target all valid
-                components on page. Defaults to `None`.
-            inputs: Dict mapping action function names with their inputs e.g.
-                inputs = {'filters': [], 'parameters': ['gdpPercap'], 'filter_interaction': [], 'theme_selector': True}
-
-        Returns:
-            Dict mapping target component ids to modified charts/components e.g. {'my_scatter': Figure({})}
-        """
         # TODO NOW: comment about how filters, paramters, filter_interaction currently ignored but will be used in
         #  future when do pattern matching callback.
-
         return _get_modified_page_figures(
             ctds_filter=ctx.args_grouping["external"]["filters"],
             ctds_filter_interaction=ctx.args_grouping["external"]["filter_interaction"],
             ctds_parameter=ctx.args_grouping["external"]["parameters"],
             targets=self.targets,
         )
-
-    function._function = function
 
     # update_figures
     # def _post_init(self):
